refactor: move find_config function away to core.common.file
