+++ deeppavlov/models/morpho_tagger/__main__.py
@@ -1,7 +1,7 @@
 import argparse
 
+from deeppavlov.core.common.file import find_config
 from deeppavlov.models.morpho_tagger.common import predict_with_model
-from deeppavlov.deep import find_config
 from deeppavlov.download import deep_download
 
 parser = argparse.ArgumentParser()
